requirements1.txt
astroid==2.2.5
beautifulsoup4==4.6.0
bootstrap4==0.1.0
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.4
django-bootstrap4==0.0.7
django-debug-toolbar==2.0
django-modelcluster==4.4
django-qr-code==1.0.0
django-taggit==0.23.0
django-treebeard==4.3
django-wkhtmltopdf==3.2.0
djangorestframework==3.9.2
draftjs-exporter==2.1.5
html5lib==1.0.1
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
Pyphen==0.9.5
pytz==2018.7
qrcode==6.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
typed-ast==1.3.1
Unidecode==1.0.23
urllib3==1.24.2
virtualenv==16.7.4
wagtail==2.4
WeasyPrint==47
webencodings==0.5.1
Willow==1.1
wincertstore==0.2
wrapt==1.11.1
